gateware.wishbone.spiflash: add `SPIFlashPeripheral`
Replaces `WishboneSPIFlashReader`, providing the same functionality with
additional features and improvements.

- Optional port for commanding the Flash device directly, enable it
  with `with_master=True`. A crossbar with a round-robin scheduling
  scheme is instanced to share the same `SPIPHYController`.

- Clocks can differ between the `SPIPHYController` and the
  `SPIFlashPeripheral` instances. A CDC module is created
  automatically in those cases.
